\name{highSCREEN-package}
\alias{highSCREEN-package}
\alias{highSCREEN}
\docType{package}
\title{
 High-Throughput Screening for Plate Based Essays
}
\description{
This package can be used to carry out extraction, normalization, quality control (QC), candidate hits identification and visualization for plate based essays, in drug discovery.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{highSCREEN}
\packageIndices{highSCREEN}

}
\author{
  I. D. Shterev, C. Chan and G. D. Sempowski

Maintainer: I. D. Shterev <i.shterev@duke.edu>
}
